Move tests to snapshot
