+++ pyproject.toml
@@ -95,6 +95,7 @@ sanic-testing = "^0.8"
 fastapi = {version = ">=0.65.0", optional = false}
 pytest-xprocess = "^0.18.1"
 MarkupSafe = "2.0.1"
+pytest-snapshot = "^0.8.1"
 
 [tool.poetry.extras]
 aiohttp = ["aiohttp", "pytest-aiohttp"]
@@ -110,6 +111,12 @@ chalice = ["chalice"]
 
 [tool.poetry.scripts]
 strawberry = "strawberry.cli:run"
+
+[tool.black]
+line-length = 88
+extend-exclude = '''
+tests/codegen/snapshots/
+'''
 
 [tool.isort]
 src_paths = ["strawberry", "tests", "scripts"]
