+++ tests/codegen/queries/basic.graphql
@@ -1,0 +1,5 @@
+query OperationName {
+  id
+  integer
+  anotherInteger
+}
