+++ tests/codegen/queries/custom_scalar.graphql
@@ -1,0 +1,3 @@
+query OperationName {
+  json
+}
