+++ tests/codegen/queries/enum.graphql
@@ -1,0 +1,3 @@
+query OperationName {
+  enum
+}
