+++ tests/codegen/queries/interface.graphql
@@ -1,0 +1,5 @@
+query OperationName {
+  interface {
+    id
+  }
+}
