+++ tests/codegen/queries/interface_fragments.graphql
@@ -1,0 +1,13 @@
+query OperationName {
+  interface {
+    id
+
+    ... on BlogPost {
+      title
+    }
+
+    ... on Image {
+      url
+    }
+  }
+}
