+++ tests/codegen/queries/interface_single_fragment.graphql
@@ -1,0 +1,9 @@
+query OperationName {
+  interface {
+    id
+
+    ... on BlogPost {
+      title
+    }
+  }
+}
