+++ tests/codegen/queries/multiple_types.graphql
@@ -1,0 +1,8 @@
+query OperationName {
+  person {
+    name
+  }
+  listOfPeople {
+    name
+  }
+}
