+++ tests/codegen/queries/multiple_types_optional.graphql
@@ -1,0 +1,5 @@
+query OperationName {
+  optionalPerson {
+    name
+  }
+}
