+++ tests/codegen/queries/optional_and_lists.graphql
@@ -1,0 +1,5 @@
+query OperationName {
+  optionalInt
+  listOfInt
+  listOfOptionalInt
+}
