+++ tests/codegen/snapshots/python/basic.graphql.py
@@ -1,0 +1,4 @@
+class OperationNameResult:
+    id: str
+    integer: int
+    another_integer: int
