+++ tests/codegen/snapshots/python/custom_scalar.graphql.py
@@ -1,0 +1,6 @@
+from typing import NewType
+
+JSON = NewType("JSON", str)
+
+class OperationNameResult:
+    json: JSON
