+++ tests/codegen/snapshots/python/enum.graphql.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+class Color(Enum):
+    RED = "RED"
+    GREEN = "GREEN"
+    BLUE = "BLUE"
+
+class OperationNameResult:
+    enum: Color
