+++ tests/codegen/snapshots/python/interface.graphql.py
@@ -1,0 +1,5 @@
+class OperationNameResultInterface:
+    id: str
+
+class OperationNameResult:
+    interface: OperationNameResultInterface
