+++ tests/codegen/snapshots/python/interface_single_fragment.graphql.py
@@ -1,0 +1,6 @@
+class OperationNameResultInterfaceBlogPost:
+    id: str
+    title: str
+
+class OperationNameResult:
+    interface: OperationNameResultInterfaceBlogPost
