+++ tests/codegen/snapshots/python/multiple_types.graphql.py
@@ -1,0 +1,11 @@
+from typing import List
+
+class OperationNameResultPerson:
+    name: str
+
+class OperationNameResultListOfPeople:
+    name: str
+
+class OperationNameResult:
+    person: OperationNameResultPerson
+    list_of_people: List[OperationNameResultListOfPeople]
