+++ tests/codegen/snapshots/python/multiple_types_optional.graphql.py
@@ -1,0 +1,7 @@
+from typing import Optional
+
+class OperationNameResultOptionalPerson:
+    name: str
+
+class OperationNameResult:
+    optional_person: Optional[OperationNameResultOptionalPerson]
