+++ tests/codegen/snapshots/python/optional_and_lists.graphql.py
@@ -1,0 +1,6 @@
+from typing import List, Optional
+
+class OperationNameResult:
+    optional_int: Optional[int]
+    list_of_int: List[int]
+    list_of_optional_int: List[Optional[int]]
