+++ tests/codegen/snapshots/python/union.graphql.py
@@ -1,0 +1,12 @@
+from typing import Union
+
+class OperationNameResultUnionAnimal:
+    age: int
+
+class OperationNameResultUnionPerson:
+    name: str
+
+OperationNameResultUnion = Union[OperationNameResultUnionAnimal, OperationNameResultUnionPerson]
+
+class OperationNameResult:
+    union: OperationNameResultUnion
