+++ tests/codegen/snapshots/typescript/basic.graphql.ts
@@ -1,0 +1,5 @@
+type OperationNameResult = {
+    id: string
+    integer: number
+    another_integer: number
+}
