+++ tests/codegen/snapshots/typescript/custom_scalar.graphql.ts
@@ -1,0 +1,5 @@
+type JSON = string
+
+type OperationNameResult = {
+    json: JSON
+}
