+++ tests/codegen/snapshots/typescript/enum.graphql.ts
@@ -1,0 +1,9 @@
+enum Color {
+    RED = "RED",
+    GREEN = "GREEN",
+    BLUE = "BLUE",
+}
+
+type OperationNameResult = {
+    enum: Color
+}
