+++ tests/codegen/snapshots/typescript/interface.graphql.ts
@@ -1,0 +1,7 @@
+type OperationNameResultInterface = {
+    id: string
+}
+
+type OperationNameResult = {
+    interface: OperationNameResultInterface
+}
