+++ tests/codegen/snapshots/typescript/interface_fragments.graphql.ts
@@ -1,0 +1,15 @@
+type OperationNameResultInterfaceBlogPost = {
+    id: string
+    title: string
+}
+
+type OperationNameResultInterfaceImage = {
+    id: string
+    url: string
+}
+
+type OperationNameResultInterface = OperationNameResultInterfaceBlogPost | OperationNameResultInterfaceImage
+
+type OperationNameResult = {
+    interface: OperationNameResultInterface
+}
