+++ tests/codegen/snapshots/typescript/interface_single_fragment.graphql.ts
@@ -1,0 +1,8 @@
+type OperationNameResultInterfaceBlogPost = {
+    id: string
+    title: string
+}
+
+type OperationNameResult = {
+    interface: OperationNameResultInterfaceBlogPost
+}
