+++ tests/codegen/snapshots/typescript/multiple_types.graphql.ts
@@ -1,0 +1,12 @@
+type OperationNameResultPerson = {
+    name: string
+}
+
+type OperationNameResultListOfPeople = {
+    name: string
+}
+
+type OperationNameResult = {
+    person: OperationNameResultPerson
+    list_of_people: OperationNameResultListOfPeople[]
+}
