+++ tests/codegen/snapshots/typescript/multiple_types_optional.graphql.ts
@@ -1,0 +1,7 @@
+type OperationNameResultOptionalPerson = {
+    name: string
+}
+
+type OperationNameResult = {
+    optional_person: OperationNameResultOptionalPerson | undefined
+}
