+++ tests/codegen/snapshots/typescript/optional_and_lists.graphql.ts
@@ -1,0 +1,5 @@
+type OperationNameResult = {
+    optional_int: number | undefined
+    list_of_int: number[]
+    list_of_optional_int: (number | undefined)[]
+}
