+++ tests/codegen/snapshots/typescript/union.graphql.ts
@@ -1,0 +1,13 @@
+type OperationNameResultUnionAnimal = {
+    age: number
+}
+
+type OperationNameResultUnionPerson = {
+    name: string
+}
+
+type OperationNameResultUnion = OperationNameResultUnionAnimal | OperationNameResultUnionPerson
+
+type OperationNameResult = {
+    union: OperationNameResultUnion
+}
